notes spacing as double + validation on settings form inputs

+++ src/main/java/antoine/rythm/entities/UserEntity.java
@@ -16,7 +16,7 @@ public class UserEntity {
 	@Id
 	private String email;
 
-	private int notesSpacing;
+	private double notesSpacing;
 
 	@ElementCollection
 	private Map<Integer, Character> keys;
